test: 수익률 계산 형변환에 관련된 테스트 코드 수정

+++ src/test/java/lotto/LottoTest.java
@@ -186,7 +186,7 @@ class LottoTest {
         );
 
         assertThat(lottoPrizeRecordingService.getPrizeRate(2000))
-                .isEqualTo(lottoPrizeRecordingService.getTotalPrizeMoney() / 2000);
+                .isEqualTo((double)lottoPrizeRecordingService.getTotalPrizeMoney() / 2000);
     }
 
 }
